requirements.txt
certifi==2024.12.14
geopandas
grpcio==1.69.0
grpcio-tools==1.69.0
numpy
packaging==24.2
pandas
protobuf
pyogrio
pyproj
python-dateutil
pytz
setuptools
shapely
six
tzdata
